+++ apps/demo-host/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @varshylinc/demo-host
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @varshylinc/onboarding-consent-engine@0.2.3
+  - @varshylinc/auth-social@0.2.3
+  - @varshylinc/mobile-payments@0.1.4
+  - @varshylinc/team-management@0.2.3
+
 ## 0.0.4
 
 ### Patch Changes